src/test/java/seedu/address/logic/commands/MapCommandTest.java
//@@author jshoung
package seedu.address.logic.commands;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertTrue;
import static seedu.address.logic.commands.MapCommand.SHOWING_MAP_MESSAGE;

import org.junit.Rule;
import org.junit.Test;

import seedu.address.commons.events.ui.ShowMapRequestEvent;
import seedu.address.ui.testutil.EventsCollectorRule;

public class MapCommandTest {
    @Rule
    public final EventsCollectorRule eventsCollectorRule = new EventsCollectorRule();

    @Test
    public void execute_map_success() {
        CommandResult result = new MapCommand().execute();
        assertEquals(SHOWING_MAP_MESSAGE, result.feedbackToUser);
        assertTrue(eventsCollectorRule.eventsCollector.getMostRecent() instanceof ShowMapRequestEvent);
        assertTrue(eventsCollectorRule.eventsCollector.getSize() == 1);
    }
}
